complete user report database and user cases part. will carry on tests later.
