+++ src/main/java/Reporter_Database/GroupRepoInt.java
@@ -1,16 +1,7 @@
 package Reporter_Database;
 
 public interface GroupRepoInt {
-
-
         /*add blacklist users to database
          */
     void adduserModel(RepoMessageModel userMessageModel);
-
-
-
-
-
-
-
 }
