+++ src/main/java/Reporter_entities/Blocker.java
@@ -1,4 +1,4 @@
-package entities;
+package Reporter_entities;
 
 public class Blocker {
     private String messageId;
