+++ src/main/java/Reporter_frameworks_and_drivers/Failed.java
@@ -1,4 +1,4 @@
-package frameworks_and_drivers;
+package Reporter_frameworks_and_drivers;
 
 public class Failed extends RuntimeException {
     public Failed(String error) {
