+++ src/main/java/add_blocked_Users/application_business_rules/AddBlockedUserRequestModel.java
@@ -22,5 +22,4 @@ public class AddBlockedUserRequestModel {
     public String getBlockedName() {
         return blockedName;
     }
-
 }
