+++ src/main/java/add_blocked_Users/application_business_rules/AddBlockedUserResponseModel.java
@@ -26,5 +26,4 @@ public class AddBlockedUserResponseModel {
     public void setBlockedList(String list) {
         this.blockedList = list;
     }
-
 }
